pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
asyncpg = "^0.30.0"
alembic = "^1.16.4"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
